Moved D and P vectors from LinkState to Router.
Migrated the storage of the D, P, and N\' ArrayLists from LinkState to Router as they truly belong in Router, not in LinkState. LinkState doesn't need to have access to them as the Router is the only object that manipulates them.  StepPrinter needs access to them, but only read access, so StepPrinter has been updated to access these values from the Router instead of the LinkState object.